Display errors on voting

+++ app/controllers/user/hub/ratings_controller.rb
@@ -1,6 +1,8 @@
 class User::Hub::RatingsController < UserApplicationController
 
   def destroy
+    head :unauthorized and return unless current_user.present?
+
     rating = Rating.find_by(user: current_user, post: Post.find(params[:post_id])) rescue nil
     head :not_found and return unless rating
     rating.destroy!
@@ -8,13 +10,22 @@ class User::Hub::RatingsController < UserApplicationController
   end
 
   def create
+    unless current_user.present?
+      render json: { message: "You must be logged in to vote." }, status: :unauthorized and return
+    end
+
     unless params[:vote].present? and %w(upvote downvote).include?(params[:vote].to_s)
       head :bad_request and return
     end
 
     symbol = { "upvote" => :up_vote, "downvote" => :down_vote }[params[:vote]]
 
-    self.destroy
+    rating = Rating.find_by(user: current_user, post: Post.find(params[:post_id])) rescue nil
+    if rating.present?
+      rating.vote = symbol
+      rating.save!
+      render json: rating, status: :ok and return
+    end
 
     rating = Rating.create!(user: current_user, post: Post.find(params[:post_id]), vote: symbol)
     render json: rating, status: :created
